Add anonymized site identifier to all tables

Add anonymized site identifier to all tables

Usage

attach_anon_id(all_sites_tbl, tbls_to_anon)

Arguments

all_sites_tbl

a tbl that contains a column called site with each of the sites that you want to mask in all the other tables

tbls_to_anon

a NAMED list of all of the tables that you want to add the masked site identifier column to

Value

each of the original tables in tbls_to_anon with the original columns plus a column called site_anon with a masked identifier that is consistent across all of the tables in tbls_to_anon