Facts Over Time

This function will compute the number of rows, patients, and (optionally) visits associated with the fact of interest within a specified time period. The user will supply the end points of the time span (i.e. January 2009 - January 2024) and the time period they wish to divide it by (i.e. month, year).

Usage

check_fot(
  fot_tbl,
  omop_or_pcornet = "omop",
  compute_method = "loop",
  time_span = list("2009-01-01", today()),
  time_period = "month",
  lookback_interval = 1,
  check_string = "fot",
  visits_only = FALSE,
  distinct_visits = TRUE
)

Arguments

fot_tbl

a table with information describing the fact tables that should be examined; see ?fot_input_omop or ?fot_input_pcornet for details

omop_or_pcornet

string indicating the CDM format of the data; defaults to omop

compute_method

a string input of either loop or group; controls whether the check is executed by looping through each time period or grouping by a date field

time_span

a list that contains the start date and end date of the time span (i.e. list(2009-01-01, 2015-01-01))

time_period

string indicating the length of time that the time span should be divided into (i.e. months, years)

lookback_interval

the number of time periods (defined in time_period) to look back in each interval (i.e. 1 year, 3 months); defaults to 1

check_string

the abbreviated name of the check; defaults to fot

visits_only

if TRUE, counts ONLY distinct visits and not patients or rows

distinct_visits

if TRUE, counts distinct visits as well as total counts and total patients

Value

a dataframe with one row for each time period within the specified time span for each check;

if visits_only = TRUE, will produce only counts of visits for the check + time period

if visits_only = FALSE and distinct_visits = TRUE, will produce counts of patients, rows, and visits for the check + time period

if visits_only = FALSE and distinct_visits = FALSE, will produce counts of patients and rows for the check + time period

Examples

# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::fot_input_omop
#> # A tibble: 22 × 6
#>    check_id  check_description              schema table date_field filter_logic
#>    <chr>     <chr>                          <chr>  <chr> <chr>      <chr>       
#>  1 vi        all visits                     cdm    visi… visit_sta… NA          
#>  2 vo_office outpatient visits              cdm    visi… visit_sta… visit_conce…
#>  3 vo_labs   outpatient lab visits          cdm    visi… visit_sta… visit_conce…
#>  4 vo_th     telehealth visits              cdm    visi… visit_sta… visit_conce…
#>  5 vo_oa     other ambulatory visits        cdm    visi… visit_sta… visit_conce…
#>  6 vip       inpatient visits (9201 or 200… cdm    visi… visit_sta… visit_conce…
#>  7 ved       emergency department visits    cdm    visi… visit_sta… visit_conce…
#>  8 vob       observation visits             cdm    visi… visit_sta… visit_conce…
#>  9 co        all conditions                 cdm    cond… condition… NA          
#> 10 dp        prescription drugs             cdm    drug… drug_expo… drug_type_c…
#> # ℹ 12 more rows
ndq::fot_input_pcornet
#> # A tibble: 21 × 6
#>    check_id  check_description              schema table date_field filter_logic
#>    <chr>     <chr>                          <chr>  <chr> <chr>      <chr>       
#>  1 vi        all visits                     cdm    enco… admit_date NA          
#>  2 vo_office outpatient visits              cdm    enco… admit_date enc_type ==…
#>  3 vo_th     telehealth visits              cdm    enco… admit_date enc_type ==…
#>  4 vo_oa     other ambulatory visits        cdm    enco… admit_date enc_type ==…
#>  5 vip       inpatient visits (9201 or 200… cdm    enco… admit_date enc_type %i…
#>  6 ved       emergency department visits    cdm    enco… admit_date enc_type ==…
#>  7 vob       observation visits             cdm    enco… admit_date enc_type ==…
#>  8 co        all diagnoses                  cdm    diag… admit_date NA          
#>  9 dp        prescription drugs             cdm    pres… rx_start_… NA          
#> 10 ip        inpatient administration drugs cdm    med_… medadmin_… NA          
#> # ℹ 11 more rows

# Use this as your input to the FOT function
## This check can be executed for different time period lengths, like...
### Yearly
if (FALSE) { # \dontrun{
my_fot_rslt <- check_fot(fot_tbl = ndq::fot_input_omop,
                         omop_or_pcornet = 'omop',
                         compute_method = 'loop', # use 'group' for high performant DBMSs
                         time_span = list('2015-01-01', '2025-01-01'),
                         time_period = 'year',
                         lookback_interval = 1,
                         check_string = 'fot')
} # }

### Monthly
if (FALSE) { # \dontrun{
my_fot_rslt <- check_fot(fot_tbl = ndq::fot_input_omop,
                         omop_or_pcornet = 'omop',
                         compute_method = 'loop', # use 'group' for high performant DBMSs
                         time_span = list('2015-01-01', '2025-01-01'),
                         time_period = 'month',
                         lookback_interval = 1,
                         check_string = 'fot')
} # }

### Quarterly
if (FALSE) { # \dontrun{
my_fot_rslt <- check_fot(fot_tbl = ndq::fot_input_omop,
                         omop_or_pcornet = 'omop',
                         compute_method = 'loop', # use 'group' for high performant DBMSs
                         time_span = list('2015-01-01', '2025-01-01'),
                         time_period = 'month',
                         lookback_interval = 3,
                         check_string = 'fot')
} # }