Missing Field: Visit ID

This function will check to see if the visit_occurrence_id/encounterid in a given fact table also exists in the visit_occurrence/encoubter table and identify cases where the visit_occurrence_id/encounterid is missing entirely (NULL). There may be cases where this is expected (for example, immunizations imported from an external registry) but generally the visit_occurrence_id/encounterid should be populated and exist as a primary key in the visit_occurrence/encounter table.

Usage

check_mf_visitid(
  mf_tbl,
  omop_or_pcornet = "omop",
  visit_tbl = cdm_tbl("visit_occurrence"),
  check_string = "mf_visitid"
)

Arguments

mf_tbl

a table with information about the tables that should be cross checked against the visit_tbl to ensure the visit_occurrence_id / encounterid exists as a primary key

omop_or_pcornet

string indicating the CDM format of the data; defaults to omop

visit_tbl

the CDM visit_occurrence / encounter table that contains the visit ID primary key

check_string

an abbreviated identifier to identify all output from this module defaults to mf_visitid

Value

a table summarizing the total number of visits in the fact table, the number of NULL visits, and the number of visits that cannot link back to the visit table. has columns: measure, total_visits, missing_visits_total, missing_visits_distinct, visit_na, total_id, check_name, database_version,site

Examples

# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::mf_input_omop
#> # A tibble: 4 × 5
#>   check_id check_description                 schema table           filter_logic
#>   <chr>    <chr>                             <chr>  <chr>           <chr>       
#> 1 coied    conditions excluding problem list cdm    condition_occu… !condition_…
#> 2 dr       all drugs                         cdm    drug_exposure   NA          
#> 3 dp       prescription or inpatient drugs   cdm    drug_exposure   !drug_type_…
#> 4 pr       all procedures                    cdm    procedure_occu… NA          
ndq::mf_input_pcornet
#> # A tibble: 4 × 5
#>   check_id check_description                 schema table       filter_logic
#>   <chr>    <chr>                             <chr>  <chr>       <lgl>       
#> 1 coied    conditions excluding problem list cdm    diagnosis   NA          
#> 2 admin    inpatient administrations         cdm    med_admin   NA          
#> 3 rx       prescription                      cdm    prescribing NA          
#> 4 pr       all procedures                    cdm    procedures  NA          

# Use this as your input to the MF: Visit ID function
if (FALSE) { # \dontrun{
my_mf_rslt <- check_mf(mf_tbl = ndq::mf_input_omop,
                       omop_or_pcornet = 'omop',
                       visit_tbl = cdm_tbl('visit_occurrence'), ## table with visit ID primary keys
                       check_string = 'mf_visitid')
} # }