Valueset Conformance

This function will intake a limited valueset that is expected to make up the entire contents of a field (minus the specified null_values) and identify if any non-permitted values exist in the field (and how often).

Usage

check_vs(
  vs_tbl,
  omop_or_pcornet = "omop",
  check_string = "vs",
  concept_tbl = vocabulary_tbl("concept"),
  null_values = c(44814650L, 0L, 44814653L, 44814649L)
)

Arguments

vs_tbl

a table with the table, field, and valueset information for each check

omop_or_pcornet

string indicating the CDM format of the data; defaults to omop

check_string

an abbreviated identifier to identify all output from this module defaults to vs

concept_tbl

a vocabulary table, like the OMOP concept table, with at least the concept column of interest (concept_id or concept_code), the concept name, and the vocabulary id

null_values

a vector of NULL values (or other values that are not part of the valueset but are broadly accepted) that should be excluded when identifying non-valueset concepts

Value

a dataframe with summary information about each value that does not appear in the valueset, or a row with dummy information if no violations are identified

Examples

# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::vs_input_omop
#> # A tibble: 2 × 6
#>   check_id    schema table  valueset_name      concept_field        filter_logic
#>   <chr>       <chr>  <chr>  <chr>              <chr>                <lgl>       
#> 1 pd_race_cid cdm    person valueset_race      race_concept_id      NA          
#> 2 pd_eth_cid  cdm    person valueset_ethnicity ethnicity_concept_id NA          
ndq::vs_input_pcornet
#> # A tibble: 2 × 7
#>   check_id    schema table       valueset_name    concept_field vocabulary_field
#>   <chr>       <chr>  <chr>       <chr>            <chr>         <lgl>           
#> 1 pd_race_cid cdm    demographic valueset_race    race          NA              
#> 2 pd_eth_cid  cdm    demographic valueset_ethnic… hispanic      NA              
#> # ℹ 1 more variable: filter_logic <lgl>

# Use this as your input to the VS function
if (FALSE) { # \dontrun{
my_vs_rslt <- check_vs(vs_tbl = ndq::vs_input_omop,
                       omop_or_pcornet = 'omop',
                       concept_tbl = vocabulary_tbl("concept"), ## points to OHDSI concept table
                       null_values = c(44814650L,0L,
                                       44814653L,44814649L), # ignored illegal values
                       check_string = 'vs')
} # }