Unmapped Concepts – Processing

Intakes the output of check_uc in order to apply additional processing. This includes either adding proportions (for by_year output) or computing overall totals across all sites included in the input (for not by_year output)

Usage

process_uc(uc_results, rslt_source = "remote", csv_rslt_path = NULL)

Arguments

uc_results

the output of check_uc

rslt_source

the location of the results. acceptable values are local (stored as a dataframe in the R environment), csv (stored as CSV files), or remote (stored on a remote DBMS); defaults to remote

csv_rslt_path

if the results have been stored as CSV files, the path to the location of these files. If the results are local or remote, leave NULL

Value

a table with either an additional column with unmapped proportions (by_year) or with additional rows that include total unmapped counts/proportions across all sites (not by_year)

Examples

# This function should be run after check_uc has been executed for all
# network institutions and results have been combined into a common table

# Once the labels have been applied, the function can be executed
## When results are kept locally:
if (FALSE) { # \dontrun{
my_uc_process <- process_uc(uc_results = my_uc_rslts,
                            rslt_source = 'local')
} # }

## When results are kept in CSV files:
if (FALSE) { # \dontrun{
my_uc_process <- process_uc(uc_results = 'my_uc_rslts',
                            rslt_source = 'csv',
                            csv_rslt_path = 'path/to/my/results')
} # }

## When results are kept on a remote database:
if (FALSE) { # \dontrun{
my_uc_process <- process_uc(uc_results = 'my_uc_rslts',
                            rslt_source = 'remote')
} # }