Add summary columns for a large N of sites

Add summary columns for a large N of sites

Usage

summarize_large_n(
  dq_output,
  num_col,
  grp_vars,
  check_string,
  time = FALSE,
  shape
)

Arguments

dq_output

the output table from processing to which the columns should be added

num_col

the numeric column to be used to compute the statistics

grp_vars

grouping variables for the computation

check_string

string indicating the check type

time

a logical indicating whether the check is time dependent

shape

'long' or 'wide' indicating whether summary statistics should be separate columns (wide) or assigned to separate rows, with the name of the summary statistic in the site column (long)

Value

a dataframe with summary statistics for the numeric column provided. contains all columns from dq_output plus:

    if time = `TRUE`, the mean and median are returned.

    if time = `FALSE`, the max, min, q1, q3, mean, and median are returned

    These values are treated as "sites" to assist with Shiny implementation